support for late key manager setup and remove password from onboarding
