+++ interface/app/onboarding/index.tsx
@@ -1,6 +1,5 @@
 import { Navigate, RouteObject } from 'react-router';
 import CreatingLibrary from './creating-library';
-import MasterPassword from './master-password';
 import NewLibrary from './new-library';
 import Privacy from './privacy';
 import Start from './start';
@@ -17,10 +16,6 @@ export default [
 	{
 		element: <NewLibrary />,
 		path: 'new-library'
-	},
-	{
-		element: <MasterPassword />,
-		path: 'master-password'
 	},
 	{
 		element: <Privacy />,
