+++ packages/client/src/core.ts
@@ -11,6 +11,7 @@ export type Procedures = {
         { key: "keys.getKey", input: LibraryArgs<string>, result: string } | 
         { key: "keys.getSecretKey", input: LibraryArgs<null>, result: string | null } | 
         { key: "keys.isKeyManagerUnlocking", input: LibraryArgs<null>, result: boolean | null } | 
+        { key: "keys.isSetup", input: LibraryArgs<null>, result: boolean } | 
         { key: "keys.isUnlocked", input: LibraryArgs<null>, result: boolean } | 
         { key: "keys.list", input: LibraryArgs<null>, result: StoredKey[] } | 
         { key: "keys.listMounted", input: LibraryArgs<null>, result: string[] } | 
@@ -53,6 +54,7 @@ export type Procedures = {
         { key: "keys.mount", input: LibraryArgs<string>, result: null } | 
         { key: "keys.restoreKeystore", input: LibraryArgs<RestoreBackupArgs>, result: number } | 
         { key: "keys.setDefault", input: LibraryArgs<string>, result: null } | 
+        { key: "keys.setup", input: LibraryArgs<OnboardingConfig>, result: null } | 
         { key: "keys.syncKeyToLibrary", input: LibraryArgs<string>, result: null } | 
         { key: "keys.unlockKeyManager", input: LibraryArgs<UnlockKeyManagerArgs>, result: null } | 
         { key: "keys.unmount", input: LibraryArgs<string>, result: null } | 
@@ -89,8 +91,6 @@ export type Procedures = {
  */
 export type Algorithm = "XChaCha20Poly1305" | "Aes256Gcm"
 
-export type AuthOption = { type: "Password", value: string } | { type: "TokenizedPassword", value: string }
-
 export type AutomountUpdateArgs = { uuid: string, status: boolean }
 
 export type BuildInfo = { version: string, commit: string }
@@ -99,7 +99,7 @@ export type CRDTOperation = { node: string, timestamp: number, id: string, typ: 
 
 export type CRDTOperationType = SharedOperation | RelationOperation | OwnedOperation
 
-export type CreateLibraryArgs = { name: string, auth: AuthOption, algorithm: Algorithm, hashing_algorithm: HashingAlgorithm }
+export type CreateLibraryArgs = { name: string }
 
 export type EditLibraryArgs = { id: string, name: string | null, description: string | null }
 
@@ -221,6 +221,8 @@ export type Nonce = { XChaCha20Poly1305: number[] } | { Aes256Gcm: number[] }
 export type Object = { id: number, pub_id: number[], kind: number, key_id: number | null, hidden: boolean, favorite: boolean, important: boolean, has_thumbnail: boolean, has_thumbstrip: boolean, has_video_preview: boolean, ipfs_id: string | null, note: string | null, date_created: string }
 
 export type ObjectValidatorArgs = { id: number, path: string }
+
+export type OnboardingConfig = { password: string, algorithm: Algorithm, hashing_algorithm: HashingAlgorithm }
 
 /**
  *  Represents the operating system which the remote peer is running.
