+++ packages/client/src/stores/onboardingStore.ts
@@ -15,10 +15,6 @@ const onboardingStoreDefaults = {
 	newLibraryName: '',
 	unlockedScreens: ['start'],
 	lastActiveScreen: null as string | null,
-	shouldEncryptLibrary: false,
-	algorithm: 'XChaCha20Poly1305' as Algorithm,
-	hashingAlgorithm: 'Argon2id-s' as HashingAlgoSlug,
-	passwordSetToken: null as string | null,
 	shareTelemetry: true,
 	useCases: [] as UseCase[],
 	grantedFullDiskAccess: false
